+++ fn.py
@@ -2,10 +2,12 @@ import bpy, re
 from mathutils import Vector
 
 
-## Some function are taken from Import image as plane addon (with some tweak)
-## credits goes to Florian Meyer (tstscr), mont29, matali, Ted Schundler (SpkyElctrc)
+
 # import importlib
 # iap = importlib.import_module('io_import_images_as_planes')
+
+## //--- Function are taken from Import image as plane addon (with some tweaks)
+## credits to Florian Meyer (tstscr), mont29, matali, Ted Schundler (SpkyElctrc)
 
 from collections import namedtuple
 ImageSpec = namedtuple(
@@ -151,17 +153,18 @@ def get_shadeless_node(dest_node_tree):
 
     return group_node
 
-#### can load from IAP ///
+#### above func can be loaded from IAP via import  (below are modified) ///
 
 def apply_texture_options(texture, img_spec):
         image_user = texture.image_user
-        # image_user.use_auto_refresh = False # self.use_auto_refresh
-        # image_user.frame_start = img_spec.frame_start
-        # image_user.frame_offset = img_spec.frame_offset
-        # image_user.frame_duration = img_spec.frame_duration
+        # image_user.use_auto_refresh = True # self.use_auto_refresh
+        image_user.frame_start = img_spec.frame_start
+        image_user.frame_offset = img_spec.frame_offset
+        image_user.frame_duration = img_spec.frame_duration
 
         # Image sequences need auto refresh to display reliably
-        if img_spec.image.source == 'SEQUENCE':
+        # if img_spec.image.source == 'SEQUENCE':
+        if img_spec.image.source in {'SEQUENCE', 'MOVIE'}:
             image_user.use_auto_refresh = True
 
         texture.extension = 'CLIP'  # Default of "Repeat" can cause artifacts
@@ -227,6 +230,8 @@ def create_cycles_material(context, img_spec, shader='EMISSION', overwrite_mater
         auto_align_nodes(node_tree)
         return material
 
+ ## Import image as plane addon function ---//
+
 
 def create_image_plane(coords, name):
     '''Create an a mesh plane with a defaut UVmap from passed coordinate
@@ -274,6 +279,7 @@ def get_ref_object_space_coord(o):
 def convert_empty_image_to_mesh(context, o, name_from='IMAGE', delete_ref=True, shader='EMISSION'):
     """shader in ['EMISSION', 'PRINCIPLED', 'SHADELESS']"""
     img = o.data  
+    img_user = o.image_user
     col = o.users_collection[0]
     
     if name_from == 'IMAGE':
@@ -296,8 +302,12 @@ def convert_empty_image_to_mesh(context, o, name_from='IMAGE', delete_ref=True, 
 
     # link in same collection
     col.objects.link(plane)
+    
+    ## img_spec :: 'image', 'size', 'frame_start', 'frame_offset', 'frame_duration'
+    # if img.source == 'FILE':
+    #     img_spec = ImageSpec(img, (0,0), 0, 0, 1) # just to match IAP norm to send... (lazy mode)
 
-    img_spec = ImageSpec(img, (0,0), 0, 0, 0) # just to match IAP norm to send... (lazy mode)
+    img_spec = ImageSpec(img, (img.size[0], img.size[1]), img_user.frame_start, img_user.frame_offset, img_user.frame_duration)
     material = create_cycles_material(context, img_spec, shader=shader)
     plane.data.materials.append(material)
 
@@ -307,3 +317,229 @@ def convert_empty_image_to_mesh(context, o, name_from='IMAGE', delete_ref=True, 
 
     if delete_ref:
         bpy.data.objects.remove(o)
+
+
+#### ----
+## FROM CAM REFS
+#### ----
+
+def set_collection(ob, collection, unlink=True) :
+    ''' link an object in a collection and create it if necessary, if unlink object is removed from other collections'''
+    scn     = bpy.context.scene
+    col     = None
+    visible = False
+    linked  = False
+
+    # check if collection exist or create it
+    for c in bpy.data.collections :
+        if c.name == collection : col = c
+    if not col : col = bpy.data.collections.new(name=collection)
+
+    # link the collection to the scene's collection if necessary
+    for c in scn.collection.children :
+        if c.name == col.name : visible = True
+    if not visible : scn.collection.children.link(col)
+
+    # check if the object is already in the collection and link it if necessary
+    for o in col.objects :
+        if o == ob : linked = True
+    if not linked : col.objects.link(ob)
+
+    # remove object from scene's collection
+    for o in scn.collection.objects :
+        if o == ob : scn.collection.objects.unlink(ob)
+
+    # if unlink flag we remove the object from other collections
+    if unlink :
+        for c in ob.users_collection :
+            if c.name != collection : c.objects.unlink(ob)
+
+def setup_transform_from_dist(cam, plane, distance):
+    from math import tan
+    plane.parent = cam
+    
+    FOV = cam.data.angle
+    shift_x = cam.data.shift_x
+    shift_y = cam.data.shift_y
+
+    plane.location.x = tan(FOV/2) * distance*2 * shift_x
+    plane.location.y = tan(FOV/2) * distance*2 * shift_y
+    plane.location.z = -distance
+    
+    for axis in range(2):
+        plane.scale[axis] = tan(FOV/2) * distance*2# * scale/100.0
+    
+    # unparent from cam after ?
+
+def create_plane_driver(cam, plane, distance=None):
+    '''Setup custom prop and drivers on object transform to fit in cam frustum
+    Code from les Fées spéciale
+    https://gitlab.com/lfs.coop/blender/camera-plane
+    '''
+
+    if not plane.get("camera_plane_distance"):
+        plane["camera_plane_distance"] = 10
+    if not plane.get("camera_plane_scale"):
+        plane["camera_plane_scale"] = 100
+    
+    plane.parent = cam
+    # plane.show_wire = True
+    # plane.matrix_world = cam.matrix_world # Already created with init transform
+    plane.lock_location = (True,)*3
+    plane.lock_rotation = (True,)*3
+    plane.lock_scale =    (True,)*3
+
+    ## LOC X AND Y (shift) ##
+    for axis in range(2):
+        driver = plane.driver_add('location', axis)
+
+        # Driver type
+        driver.driver.type = 'SCRIPTED'
+
+        # Variable DISTANCE
+        var = driver.driver.variables.new()
+        var.name = "distance"
+        var.type = 'SINGLE_PROP'
+        var.targets[0].id = plane
+        var.targets[0].data_path = '["camera_plane_distance"]'
+
+        # Variable FOV
+        var = driver.driver.variables.new()
+        var.name = "FOV"
+        var.type = 'SINGLE_PROP'
+        var.targets[0].id_type = "OBJECT"
+        var.targets[0].id = cam
+        var.targets[0].data_path = 'data.angle'
+
+        # Variable scale
+        var = driver.driver.variables.new()
+        var.name = "shift"
+        var.type = 'SINGLE_PROP'
+        var.targets[0].id = cam
+        var.targets[0].data_path = 'data.shift_' + ('x' if axis == 0 else 'y')
+
+        # Expression
+        driver.driver.expression = \
+            "tan(FOV/2) * distance*2 * shift"
+
+    ## DISTANCE ##
+    driver = plane.driver_add('location', 2)
+    # Driver type
+    driver.driver.type = 'SCRIPTED'
+    # Variable
+    var = driver.driver.variables.new()
+    var.name = "distance"
+    var.type = 'SINGLE_PROP'
+    var.targets[0].id = plane
+    var.targets[0].data_path = '["camera_plane_distance"]'
+
+    # Expression
+    driver.driver.expression = "-distance"
+
+    ## SCALE X AND Y ##
+    for axis in range(2):
+        driver = plane.driver_add('scale', axis)
+
+        # Driver type
+        driver.driver.type = 'SCRIPTED'
+
+        # Variable DISTANCE
+        var = driver.driver.variables.new()
+        var.name = "distance"
+        var.type = 'SINGLE_PROP'
+        var.targets[0].id = plane
+        var.targets[0].data_path = '["camera_plane_distance"]'
+
+        # Variable FOV
+        var = driver.driver.variables.new()
+        var.name = "FOV"
+        var.type = 'SINGLE_PROP'
+        var.targets[0].id_type = "OBJECT"
+        var.targets[0].id = cam
+        var.targets[0].data_path = 'data.angle'
+
+        # Variable scale
+        var = driver.driver.variables.new()
+        var.name = "scale"
+        var.type = 'SINGLE_PROP'
+        var.targets[0].id = plane
+        var.targets[0].data_path = '["camera_plane_scale"]'
+
+        # Expression
+        driver.driver.expression = \
+            "tan(FOV/2) * distance*2 * scale/100.0"
+    
+    # Set init distance if given
+    if distance:
+        plane["camera_plane_distance"] = distance
+    
+
+def convert_cam_bg_image_to_mesh(context, shader='EMISSION', distance=10, use_driver=False, post_state='HIDE'):
+    cam = context.object
+    if not cam or cam.type != 'CAMERA':
+        # fallback to scene active camera if no cam selected
+        cam = context.scene.camera
+    if not cam:
+        return
+
+    offset = 0
+    for bg in reversed(cam.data.background_images):
+        if not bg.show_background_image:
+            continue
+
+        img = bg.image
+        if not img:
+            continue
+        
+        name = img.name
+
+        # create plane coords (just make width equal to 1)
+        res_x, res_y = img.size
+        
+        # scaling = 1 / max(res_x, res_y)
+        scaling = 1 / res_x
+
+        # 3----2
+        # |    |
+        # 0----1 
+
+        corners = [
+            Vector((0,0)),
+            Vector((res_x, 0)),
+            Vector((0, res_y)),
+            Vector((res_x, res_y)),
+            ]
+
+        width_based_corners = []
+        for co in corners:
+            nco_x = (co.x + (-0.5 * res_x)) * scaling
+            nco_y = (co.y + (-0.5 * res_y)) * scaling
+            width_based_corners.append(Vector((nco_x, nco_y, 0)))
+
+        plane = create_image_plane(width_based_corners, name)
+        
+        #if bg.source == 'IMAGE' and bg.image:
+
+        ## 'image', 'size', 'frame_start', 'frame_offset', 'frame_duration'
+        img_spec = ImageSpec(img, (res_x, res_y), bg.image_user.frame_start, bg.image_user.frame_offset, bg.image_user.frame_duration)
+        material = create_cycles_material(context, img_spec, shader=shader)
+        plane.data.materials.append(material)
+
+
+        # context.scene.collection.objects.link(plane) # scene collection
+        set_collection(plane, "Background")
+
+        # create ctrl driver
+        if use_driver:
+            create_plane_driver(cam, plane, distance + offset)
+        else:
+            setup_transform_from_dist(cam, plane, distance + offset)
+        
+
+        if post_state == 'HIDE':
+            bg.show_background_image = False
+        elif post_state == 'DELETE':
+            cam.data.background_images.remove(bg)
+        
+        offset += 2 # offset when mulitple ref to generate
+        print(f'Generated image plane {plane.name}')