+++ CHANGELOG.md
@@ -1,14 +1,27 @@
 # Changelog
 
+<!-- 
+TODO:
+add menu entry to object (with poll ?) for empty object selection
+add panel in bg image reference for conversion
+ -->
+
+0.3.0
+
+- feat: add plane generation from camera background image (driver method taken from [LFS camera_plane](https://gitlab.com/lfs.coop/blender/camera-plane/-/blob/master/camera_plane.py))
+- ui: add menu entry in `View 3D > Object > Convert` for object conversion
+- ui: add button in `Propeties > cam data > background image` for bg image conversion
+- doc: update to reflect changes
+
 0.2.0
 
 - feat: option to get name from orignal image or from empty object
 - fix: bug with linking
 - doc: fic readme link and description
 - code: cleaner:
-    - copied function locally instead of importing from IAP
-    - add bug report url
-    - add git-ignore
+  - copied function locally instead of importing from IAP
+  - add bug report url
+  - add git-ignore
 
 0.1.0